Modifying Field squares to be more efficient
- After some profiling, it appears that the program spends a fair
bit of time assigning balls to squares for collision detection.
- This commit is partway through an attempt to rewrite the ball
assignment algorithm to be more efficient, removing linear searches
and trying to adjust it to allow squares smaller than the half the
radius of the ball. Hopefully when these changes are finished, the
program will run more efficiently. Time will tell.